Created AreaController and the index.gsp for serviced area list.
The AreaController will be used to determine if a visitor is attempting to use a tenant that our clone services. For example, if a person visits sanantonio.gcraigslist.com and the city is not in our City table, they will be given the general area index and the request will be logged to guage interest for other cities. If they visit odessa.gcraigslist.com, they will get the content specific to that area.

I don't own gcraigslist.com*